set correct device for local rank
