+++ src/distribution.cpp
@@ -1,32 +1,51 @@
 #include "distribution.hpp"
+#include "lib_dispatch.hpp"
 #include <cmath>
 #include <mpi.h>
 #include <numeric>
 
-
-
-std::array<int, 2> initialize_distribution() {
+std::array<int, 2> initialize_distribution()
+{
 #ifdef ASGARD_USE_MPI
-    auto status = MPI_Init(NULL, NULL);
-    assert(status == 0);
-    int num_ranks;
-    status = MPI_Comm_size(MPI_COMM_WORLD, &num_ranks);
-    assert(status == 0);
-    int my_rank;
-    status = MPI_Comm_rank(MPI_COMM_WORLD, &my_rank);
-    assert(status == 0);
-    return {my_rank, num_ranks};
+  auto status = MPI_Init(NULL, NULL);
+  assert(status == 0);
+  int num_ranks;
+  status = MPI_Comm_size(MPI_COMM_WORLD, &num_ranks);
+  assert(status == 0);
+  int my_rank;
+  status = MPI_Comm_rank(MPI_COMM_WORLD, &my_rank);
+  assert(status == 0);
+  initialize_libraries(get_local_rank());
+  return {my_rank, num_ranks};
 #endif
-    return {0, 1};
+  return {0, 1};
 }
 
-void finalize_distribution() {
+void finalize_distribution()
+{
 #ifdef ASGARD_USE_MPI
-	auto const status = MPI_Finalize();
-	assert(status==0);
+  auto const status = MPI_Finalize();
+  assert(status == 0);
 #endif
 }
 
+int get_local_rank()
+{
+  static int const rank = []() {
+#ifdef ASGARD_USE_MPI
+    MPI_Comm local_comm;
+    auto success = MPI_Comm_split_type(MPI_COMM_WORLD, MPI_COMM_TYPE_SHARED, 0,
+                                       MPI_INFO_NULL, &local_comm);
+    assert(success == 0);
+    int local_rank;
+    success = MPI_Comm_rank(local_comm, &local_rank);
+    assert(success == 0);
+    return local_rank;
+#endif
+    return 0;
+  }();
+  return rank;
+}
 
 // determine the side lengths that will give us the "squarest" rectangles
 // possible
