+++ src/distribution.hpp
@@ -68,6 +68,10 @@ struct element_subgrid
   int const col_stop;
 };
 
+std::array<int, 2> initialize_distribution();
+int get_local_rank();
+void finalize_distribution();
+
 struct node_out
 {
   template<class T>
@@ -77,35 +81,7 @@ struct node_out
       std::cout << val;
     return *this;
   }
-  // helper function
-  int get_local_rank()
-  {
-    static int const rank = []() {
-#ifdef ASGARD_USE_MPI
-      MPI_Comm local_comm;
-      MPI_Comm_split_type(MPI_COMM_WORLD, MPI_COMM_TYPE_SHARED, 0,
-                          MPI_INFO_NULL, &local_comm);
-      int local_rank;
-      MPI_Comm_rank(local_comm, &local_rank);
-      return local_rank;
-#endif
-      return 0;
-    }();
-    return rank;
-  }
 };
-
-inline void print(std::string const &to_print, int const my_rank,
-                  bool const master_only = true)
-{
-  if (my_rank == 0 || !master_only)
-  {
-    std::cout << to_print;
-  }
-}
-
-std::array<int, 2> initialize_distribution();
-void finalize_distribution();
 
 // given a rank, determine element subgrid
 element_subgrid
