+++ src/lib_dispatch.hpp
@@ -88,8 +88,9 @@ extern "C"
                float *b, int *ldb, int *info);
 }
 
-// -- precision/execution resource wrapper for blas --
+void initialize_libraries(int const local_rank);
 
+// -- precision/execution resource wrapper for blas --
 namespace lib_dispatch
 {
 template<typename P>
